DSM-Theme: rename to akeneoTheme
