+++ akeneo-design-system/src/theme/akeneoTheme/color.theme.ts
